\name{test.texr}
\alias{test.texr}
\title{ Runs a set of tests. }
\description{
  Runs a set of tests to check \code{texr} is working correctly.
}
\usage{
  test.texr(...)
}
\arguments{
  \item{...}{ Currently no arguments. }
}
\details{
  Runs a series of tests.
}
\seealso{ \code{\link{tex}} }
\examples{
  \dontrun{
    test.texr()
  }
}
